+++ integroBridgeService/src/main/java/com/adobe/campaign/tests/bridge/service/BridgeServiceFactory.java
@@ -31,17 +31,18 @@ public class BridgeServiceFactory {
     /**
      * Transforms the results of the JavaCall to a JSON
      * @param in_callResults A JavaCallResults object to be transformed
-     * @param in_headerValues A set of secrets that need to be check for the output
+     * @param in_secretValues A set of secrets that need to be checked for the output
      * @return The JSON representation of the Call result
      * @throws JsonProcessingException when failing to parse the JavaCallResults object
+     * @throws IBSPayloadException when the output contains secrets that have been passed
      */
-    public static String transformJavaCallResultsToJSON(JavaCallResults in_callResults, Set<String> in_headerValues)
+    public static String transformJavaCallResultsToJSON(JavaCallResults in_callResults, Set<String> in_secretValues)
             throws JsonProcessingException {
         LogManagement.logStep(LogManagement.STD_STEPS.GENERATING_RESPONSE);
 
         ObjectMapper mapper = new ObjectMapper();
         String lr_resultPayload = mapper.writeValueAsString(in_callResults);
-        if (ConfigValueHandlerIBS.HEADERS_BLOCK_OUTPUT.is("true") && in_headerValues.stream().anyMatch(h -> lr_resultPayload.contains(h))) {
+        if (ConfigValueHandlerIBS.HEADERS_BLOCK_OUTPUT.is("true") && in_secretValues.stream().anyMatch(h -> lr_resultPayload.contains(h))) {
             throw new IBSPayloadException("Your return payload contains secrets. You may consider reevaluating the headers you send. If they are not a secret, they can be put directly in the payload. Otherwise you can simply disable the "+ConfigValueHandlerIBS.HEADERS_BLOCK_OUTPUT.systemName+" option.");
         }
         return lr_resultPayload;